Phase 7.1d: Reorganize ColorIndicator to match migration pattern
- Moved plain CSS implementation to ColorIndicator.new.tsx
- Created ColorIndicator.legacy.tsx with styled wrappers
- Updated main file to export from legacy for backward compatibility
- Component already used plain CSS, just reorganizing file structure
